Fix behat: Allow switching scope from products page
