+++ features/Pim/Behat/Decorator/ContextSwitcherDecorator.php
@@ -3,7 +3,7 @@
 namespace Pim\Behat\Decorator;
 
 use Context\Spin\SpinCapableTrait;
-use Pim\Behat\Decorator\ElementDecorator;
+use Context\Spin\TimeoutException;
 
 /**
  * Decorator to add switch context feature to an element
@@ -14,7 +14,8 @@ class ContextSwitcherDecorator extends ElementDecorator
 
     protected $selectors = [
         'Locales dropdown' => '.locale-switcher',
-        'Channel dropdown' => '.scope-switcher',
+        'Channel dropdown' => '.scope-switcher, .scope-filter',
+        'Channel toggle'   => '.dropdown-toggle, .caret'
     ];
 
     /**
@@ -62,7 +63,7 @@ class ContextSwitcherDecorator extends ElementDecorator
     /**
      * @param string $scopeCode
      *
-     * @throws \Exception
+     * @throws TimeoutException
      */
     public function switchScope($scopeCode)
     {
@@ -71,12 +72,20 @@ class ContextSwitcherDecorator extends ElementDecorator
         }, 'Could not find scope switcher');
 
         $toggle = $this->spin(function () use ($dropdown) {
-            return $dropdown->find('css', '.dropdown-toggle');
+            return $dropdown->find('css', $this->selectors['Channel toggle']);
         }, 'Cannot find ".dropdown-toggle" element in scope switcher');
         $toggle->click();
 
         $option = $this->spin(function () use ($dropdown, $scopeCode) {
-            return $dropdown->find('css', sprintf('a[data-scope="%s"]', $scopeCode));
+            $item = $dropdown->find('css', sprintf('a[data-scope="%s"]', $scopeCode));
+            if (null === $item) {
+                $itemField = $this
+                    ->getBody()
+                    ->find('css', sprintf('.select-filter-widget input[title="%s"]', $scopeCode));
+                $item = (null === $itemField) ? null : $itemField->getParent();
+            }
+
+            return $item;
         }, sprintf('Could not find scope "%s" in switcher', $scopeCode));
         $option->click();
     }
